:sparkles: feat: exception handling

+++ sistask_db.py
@@ -45,11 +45,13 @@ class Task(Base):
 
             return model_pb2.TaskID(id=new_task.id)
             
-    def update(t:model_pb2.Task) -> model_pb2.TaskID:
+    def update(t:model_pb2.Task) -> model_pb2.TaskID | None:
         with Session(ENGINE) as session:
             task = session.scalars(
                 select(Task).where(Task.id == t.id)
             ).first()
+            if not task:
+                return None 
 
             task.date = t.date
             task.responsible = t.responsible
@@ -60,12 +62,13 @@ class Task(Base):
             session.commit()
             return model_pb2.TaskID(id=task.id)
 
-    def get_task(tid:model_pb2.TaskID) -> model_pb2.Task:
+    def get_task(tid:model_pb2.TaskID) -> model_pb2.Task | None:
         with Session(ENGINE) as session:
             task = session.scalars(
                 select(Task).where(Task.id == tid.id)
             ).first()
-
+            if not task:
+                return None 
             return model_pb2.Task(
                 id=task.id,
                 title=task.title,
@@ -94,22 +97,25 @@ class Task(Base):
             return list_tasks
 
                 
-    def delete(tid:model_pb2.TaskID) -> model_pb2.Void:
+    def delete(tid:model_pb2.TaskID) -> model_pb2.Void | None:
         with Session(ENGINE) as session:
             task = session.scalars(
                 select(Task).where(Task.id == tid.id)
             ).first()
-
+            if not task:
+                return None
             session.delete(task)
             session.commit()
 
             return model_pb2.Void()
 
-    def finish_task(tid:model_pb2.TaskID) -> model_pb2.TaskID:
+    def finish_task(tid:model_pb2.TaskID) -> model_pb2.TaskID | None:
         with Session(ENGINE) as session:
             task = session.scalars(
                 select(Task).where(Task.id == tid.id)
             ).first()
+            if not task:
+                return None
 
             task.is_completed = True
             session.commit()
